Add flag to print json build report (#47)

+++ src/image_1d_example.rs
@@ -1,3 +1,4 @@
+use crate::build_info;
 use crate::fft_solver::PlanType;
 use crate::util::*;
 use clap::Parser;
@@ -11,7 +12,7 @@ pub struct Args {
     /// WARNING, if this Directory
     /// already exists, current contents will be removed.
     #[arg(short, long)]
-    pub output_dir: std::path::PathBuf,
+    pub output_dir: Option<std::path::PathBuf>,
 
     /// Chunk size to use for parallelism.
     #[arg(short, long, default_value = "1000")]
@@ -30,7 +31,7 @@ pub struct Args {
     pub domain_size: usize,
 
     /// Write out image, WARNING: we do not check image size, so be reasonable.
-    #[arg(short, long)]
+    #[arg(short, long, requires("output_dir"))]
     pub write_image: bool,
 
     /// The number of threads to use.
@@ -50,7 +51,7 @@ pub struct Args {
     pub rand_init: bool,
 
     /// Write out a dot file for the ap plan
-    #[arg(long)]
+    #[arg(long, requires("output_dir"))]
     pub write_dot: bool,
 
     /// Target ratio for fft solves
@@ -64,21 +65,28 @@ pub struct Args {
     /// Generate solver only, do not solve
     #[arg(long)]
     pub gen_only: bool,
+
+    /// Print build information and quit
+    #[arg(long)]
+    pub build_info: bool,
 }
 
 impl Args {
     pub fn cli_parse(name: &str) -> (Self, PathBuf) {
-        println!("EXAMPLE: {}", name);
-        println!("GIT DESCRIBE: {}", env!("GIT_DESCRIBE"));
-        println!("GIT HASH: {}", env!("GIT_HASH"));
         let args = Args::parse();
 
-        let output_dir = args.output_dir.to_str().unwrap();
-        let _ = std::fs::remove_dir_all(output_dir);
-        std::fs::create_dir(output_dir).unwrap();
+        if args.build_info {
+            build_info::print_report(name);
+        }
 
-        let mut output_image_path = args.output_dir.clone();
-        output_image_path.push(format!("{}.png", name));
+        if let Some(output_dir) = &args.output_dir {
+            let output_dir = output_dir.to_str().unwrap();
+            let _ = std::fs::remove_dir_all(output_dir);
+            std::fs::create_dir(output_dir).unwrap();
+        }
+
+        let mut output_image_path = args.output_dir.as_ref().unwrap().clone();
+        output_image_path.push("image.png");
 
         rayon::ThreadPoolBuilder::new()
             .num_threads(args.threads)
@@ -98,6 +106,12 @@ impl Args {
 
     pub fn grid_bounds(&self) -> AABB<1> {
         AABB::new(matrix![0, self.domain_size as i32 - 1])
+    }
+
+    pub fn dot_path(&self) -> PathBuf {
+        let mut dot_path = self.output_dir.as_ref().unwrap().clone();
+        dot_path.push("plan.dot");
+        dot_path
     }
 
     pub fn save_wisdom(&self) {
